statrt to add

+++ Model/Product.cs
@@ -18,9 +18,9 @@ namespace NorthwindConsole.Model
         public int? CategoryId { get; set; }
         public string QuantityPerUnit { get; set; }
         public decimal? UnitPrice { get; set; }
-        public short? UnitsInStock { get; set; }
-        public short? UnitsOnOrder { get; set; }
-        public short? ReorderLevel { get; set; }
+        public int? UnitsInStock { get; set; }
+        public int? UnitsOnOrder { get; set; }
+        public int? ReorderLevel { get; set; }
         public bool Discontinued { get; set; }
 
         public virtual Category Category { get; set; }
